requirements.txt
# requirements.txt — Production dependencies

# ── Broker & Market Data ──────────────────────────────
alpaca-py
yfinance
pandas-datareader
edgartools
fundamentalanalysis

# ── Data & Scientific Foundation ──────────────────────
numpy
pandas
scipy
statsmodels
pyarrow
requests
httpx
python-dotenv
PyYAML

# ── Technical Analysis ────────────────────────────────
# Note: TA-Lib requires the C library to be installed on the system
TA-Lib
pandas-ta
finta

# ── Quantitative & Portfolio Analysis ─────────────────
vectorbt
Riskfolio-Lib
PyPortfolioOpt
QuantStats
hmmlearn
scikit-learn

# ── Machine Learning & NLP ────────────────────────────
transformers
torch

# ── Alerts & Notifications ────────────────────────────
# We are using SMTP-to-SMS workaround, but keeping this for future
twilio
